pyproject.toml
[project]
name = "simpleserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.2",
    "openai>=1.82.1",
    "openai-agents>=0.0.16",
    "openai-agents-mcp>=0.0.8",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
